Refactor ProfileDetails component with new structure
Renamed profile-details files to follow component naming conventions for better clarity. Introduced facade and processor services to streamline logic and configuration management. Adjusted template bindings to use the new config object.
