+++ frontend/src/app/pages/profile-details/index.ts
@@ -1,3 +1,3 @@
 // src/app/pages/profile-details/index.ts
 
-export * from "./profile-details";
+export * from "./profile-details.component";
