+++ frontend/src/app/pages/profile-details/profile-details.component.html
@@ -1,8 +1,8 @@
-<!-- src/app/pages/profile-details/profile-details.html -->
+<!-- src/app/pages/profile-details/profile-details.component.html -->
 <app-page-header
-    [title]="pageTitle"
-    [description]="pageDescription"
-    [breadcrumbs]="breadcrumbs">
+    [title]="config.pageTitle"
+    [description]="config.pageDescription"
+    [breadcrumbs]="config.breadcrumbs">
 </app-page-header>
 
 <div class="profile-details-container">
