+++ frontend/src/app/pages/profile-details/profile-details.component.scss
@@ -1,4 +1,4 @@
-// src/app/pages/profile-details/profile-details.scss
+// src/app/pages/profile-details/profile-details.component.scss
 
 .profile-details-container {
     display: grid;
